requirements.txt
django
# psycopg2-binary
requests
gunicorn


##### Rest & gRPC Framework #####
djangorestframework
djangogrpcframework
grpcio
grpcio-tools

markdown       # Markdown support for the browsable API.
django-filter

grpcWSGI


##### Opentelemetry #####
opentelemetry-api==1.5.
# opentelemetry-api==1.10a0
opentelemetry-sdk==1.5.0
# opentelemetry-sdk==1.10a0
# opentelemetry-distro==0.24b0
# opentelemetry-instrumentation
opentelemetry-instrumentation-django==0.24b0
# opentelemetry-instrumentation-wsgi==0.24b0
# opentelemetry-instrumentation-grpc==0.24b0
opentelemetry-instrumentation-requests==0.24b0
#opentelemetry-exporter-otlp==1.10a0
# opentelemetry-exporter-otlp-proto-http==1.5.0
# opentelemetry-exporter-otlp-proto-grpc==1.5.0
#opentelemetry-exporter-prometheus==1.10a0
# opentelemetry-exporter-jaeger==1.5.0
opentelemetry-exporter-gcp-trace==1.0.0
# opentelemetry-propagator-gcp==1.0.0
# psutil

##### Utils #####
django-prometheus
locust